# contracts/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ixnos-data API",
    "description": "Search Greek public procurement and spending data from ΚΗΜΔΗΣ and Διαύγεια.\n\n**Rate limits.** Without a key, 120 requests a minute per IP address. With an API key\n(free: sign in on the website, then create one on the alerts page), send it as the\n`X-Api-Key` header for 600 requests a minute. Over the limit: HTTP 429.\n\n**Bulk data.** Daily exports of every record (CSV and JSON Lines) are at /exports/.\n\n**Sources and licence.** ΚΗΜΔΗΣ and Διαύγεια open data, both CC BY 4.0: credit the\nsources when you reuse the data.",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "v1"
  },
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": [
          "Public"
        ],
        "summary": "Email a one-time sign-in link (always accepted; never reveals whether an account exists)",
        "operationId": "RequestLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/verify": {
      "post": {
        "tags": [
          "Public"
        ],
        "summary": "Exchange a sign-in link's token for a session token",
        "operationId": "VerifyLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/v1/me/digest": {
      "put": {
        "tags": [
          "Public"
        ],
        "summary": "Choose how often alerts are emailed: daily, weekly or paused",
        "operationId": "SetDigest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DigestSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/saved-searches": {
      "post": {
        "tags": [
          "Public"
        ],
        "summary": "Save a search for daily alerts",
        "operationId": "SaveSearch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/saved-searches/{id}": {
      "delete": {
        "tags": [
          "Public"
        ],
        "summary": "Stop alerts for a saved search",
        "operationId": "DeleteSavedSearch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1/me/api-keys": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "The account's active API keys (without the keys themselves)",
        "operationId": "ListApiKeys",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Public"
        ],
        "summary": "Create an API key; the key is returned once",
        "operationId": "CreateApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiKey"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/api-keys/{id}": {
      "delete": {
        "tags": [
          "Public"
        ],
        "summary": "Revoke an API key",
        "operationId": "RevokeApiKey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiKeyRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiKeyView": {
        "required": [
          "id",
          "name",
          "prefix",
          "permitsPerMinute",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "permitsPerMinute": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreatedApiKey": {
        "required": [
          "apiKey",
          "key"
        ],
        "type": "object",
        "properties": {
          "apiKey": {
            "$ref": "#/components/schemas/ApiKeyView"
          },
          "key": {
            "type": "string"
          }
        }
      },
...
      "DigestSettingsRequest": {
        "required": [
          "frequency"
        ],
        "type": "object",
        "properties": {
          "frequency": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "LoginRequest": {
        "required": [
          "email",
          "locale"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "locale": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SavedSearchRequest": {
        "required": [
          "name",
          "q",
          "kind",
          "cpv",
          "nuts",
          "minAmount",
          "maxAmount"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "q": {
            "type": [
              "null",
              "string"
            ]
          },
          "kind": {
            "type": [
              "null",
              "string"
            ]
          },
          "cpv": {
            "type": [
              "null",
              "string"
            ]
          },
          "nuts": {
            "type": [
              "null",
              "string"
            ]
          },
          "minAmount": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "maxAmount": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        }
      },
      "SavedSearchView": {
        "required": [
          "id",
          "name",
          "q",
          "kind",
          "cpv",
          "nuts",
          "minAmount",
          "maxAmount",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "q": {
            "type": [
              "null",
              "string"
            ]
          },
          "kind": {
            "type": [
              "null",
              "string"
            ]
          },
          "cpv": {
            "type": [
              "null",
              "string"
            ]
          },
          "nuts": {
            "type": [
              "null",
              "string"
            ]
          },
          "minAmount": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "maxAmount": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SessionResult": {
        "required": [
          "sessionToken",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "sessionToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "VerifyRequest": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}